Fix Cluster creation
If the cluster was not sequential from 0, i.e. used ids 1, 2, 3 as in
the quickstart, then these methods would panic. Since we're returning a
list we should just pre-allocate it as before but don't initialize the
length and just use append.

Also adds using peerLock more but this is still pretty unsafe in lots of
places. (In other ways.)